when store is deleted, the store's review&menu are deleted complete

+++ src/routes/EditTable.js
@@ -1,44 +1,44 @@
 import { authService, dbService } from "fbase";
 import React, { useEffect, useState } from "react";
 import AddOrder from "./AddOrder";
-import {Link} from "react-router-dom";
+import { Link } from "react-router-dom";
 
 const EditTable = (storeObj) => {
-    const [tableArray, setTableArray]  = useState([]);
+    const [tableArray, setTableArray] = useState([]);
     const selectedStoreID = storeObj.location.state.storeObj;
     const [idList, setIdList] = useState([]); // Tables 컬렉션의 모든 문서 값들
     const hello = "HELLO"
 
     useEffect(() => {
         dbService.collection("Tables").where
-        ("UniqueStoreId", "==", selectedStoreID).onSnapshot(snapshot => {
-            
-            const id = snapshot.docs.map(doc => (   // 테이블 문서값을 필드에 셋
-                dbService.collection("Tables").doc(doc.id).set({
-                    UniqueTableNo: doc.id
-                }, {merge : true }),
+            ("UniqueStoreId", "==", selectedStoreID).onSnapshot(snapshot => {
 
-                { // Tables 컬렉션의 문서 값만 추출
-                id: doc.id
-            }));
-            setIdList(id);
+                const id = snapshot.docs.map(doc => (   // 테이블 문서값을 필드에 셋
+                    dbService.collection("Tables").doc(doc.id).set({
+                        UniqueTableNo: doc.id
+                    }, { merge: true }),
 
-            const tableArray = snapshot.docs.map(doc => ({// Tables값 전체 가져오기
-                
-                id: doc.id,
-                ...doc.data(),
-            }));
-            setTableArray(tableArray);
-            
-        });
-        }, []
+                    { // Tables 컬렉션의 문서 값만 추출
+                        id: doc.id
+                    }));
+                setIdList(id);
+
+                const tableArray = snapshot.docs.map(doc => ({// Tables값 전체 가져오기
+
+                    id: doc.id,
+                    ...doc.data(),
+                }));
+                setTableArray(tableArray);
+
+            });
+    }, []
     );
 
     const checkLength = (event) => {
         event.preventDefault();
         console.log(idList.length)
         console.log(idList)
-        
+
     }
 
     const AddTable = async (event) => {
@@ -46,16 +46,19 @@ const EditTable = (storeObj) => {
         await dbService.collection("Tables").add({
             UniqueStoreId: selectedStoreID,
         });
-        
+
     }
 
     const DeleteTable = async (event) => {
         event.preventDefault();
-        const { target: { value } } = event;
-        dbService.collection("Tables").doc(value).delete()
+        const ok = window.confirm("해당 테이블을 삭제하시겠습니까?");
+        if (ok) {
+            const { target: { value } } = event;
+            dbService.collection("Tables").doc(value).delete();
+        }
     }
-    console.log(selectedStoreID);
-    return(
+
+    return (
         <div>
             <div>
                 <button onClick={checkLength}>idList 길이 확인</button>
@@ -66,23 +69,23 @@ const EditTable = (storeObj) => {
 
             <div>
                 {tableArray.map((obj) => (
-                        <div> 
-                            <div>
-                                {console.log(obj)}
-                                <Link to = {{
-                                    pathname:"/AddOrder",
-                                    state: {
-                                        selectedStoreID,
-                                        obj
-                                    }}
-                                }>
+                    <div>
+                        <div>
+                            <Link to={{
+                                pathname: "/AddOrder",
+                                state: {
+                                    selectedStoreID,
+                                    obj
+                                }
+                            }
+                            }>
                                 <button >테이블 고유 번호 = {obj.id}</button>
-                                </Link>
-                                <button onClick={DeleteTable} value={obj.id}> 테이블 삭제 </button>
-                            </div>
+                            </Link>
+                            <button onClick={DeleteTable} value={obj.id}> 테이블 삭제 </button>
                         </div>
-                    ))}
-            </div>   
+                    </div>
+                ))}
+            </div>
         </div>
     )
 }
